Import first draft
I'm not sure this is a good idea, but I am experimenting
with the Trello API and putting together a proof of concept.

I'm in the exploration phase, so no engineering effort is
spent (a lot of duplicated code, etc).

That's fine; there's a good chance I will throw away
all of this, so I don't want to spend extra time.